+++ byteps/common/ready_table.cc
@@ -22,12 +22,12 @@ namespace common {
 
 
 // below are methods for accessing/modifying the _ready_table
-bool ReadyTable::IsKeyReady(int key) {
+bool ReadyTable::IsKeyReady(uint64_t key) {
     std::lock_guard<std::mutex> lock(_table_mutex);
     return _ready_table[key] == (_ready_count);
 }
 
-int ReadyTable::AddReadyCount(int key) {
+int ReadyTable::AddReadyCount(uint64_t key) {
     std::lock_guard<std::mutex> lock(_table_mutex);
     BPS_CHECK_LT(_ready_table[key], _ready_count)
         << _table_name << ": "
@@ -35,7 +35,7 @@ int ReadyTable::AddReadyCount(int key) {
     return ++_ready_table[key];
 }
 
-void ReadyTable::ClearReadyCount(int key) {
+void ReadyTable::ClearReadyCount(uint64_t key) {
     std::lock_guard<std::mutex> lock(_table_mutex);
     _ready_table[key] = 0;
 }
