+++ byteps/common/scheduled_queue.h
@@ -33,7 +33,7 @@ public:
     QueueType getQueueType() { return _qt; }
     void addTask(std::shared_ptr<TensorTableEntry>);
     std::shared_ptr<TensorTableEntry> getTask();
-    std::shared_ptr<TensorTableEntry> getTask(int key);
+    std::shared_ptr<TensorTableEntry> getTask(uint64_t key);
     uint32_t pendingSize();
     void reportFinish(int size);
